Do World stuff and Start Entities

+++ src/com/mills/rendering/Colors.java
@@ -2,12 +2,12 @@ package com.mills.rendering;
 
 public class Colors {
 
-	public static int get(int color1, int color2, int color3, int color4)
+	public static long get(int color1, int color2, int color3, int color4)
 	{
 		return (get(color4) << 24) + (get(color3) << 16) + (get(color2) << 8) + get(color1);
 	}
 	
-	private static int get(int color)
+	private static long get(int color)
 	{
 		if(color < 0) return 255;
 		int r = color / 100 % 10;
